move bubble and bubbletemp to bases
