+++ datafiles/javascript/rooms/levels/Level19.js
@@ -1,5 +1,5 @@
 import Level from "./Level.js";
-import Bubble from "../../objects/Bubble.js";
+import Bubble from "../../objects/bases/Bubble.js";
 import KI2 from "../../appEtc/KI2.js";
 
 export default class Level19 extends Level {
