+++ datafiles/javascript/rooms/levels/Level21.js
@@ -1,5 +1,5 @@
 import Level from "./Level.js";
-import Bubble from "../../objects/Bubble.js";
+import Bubble from "../../objects/bases/Bubble.js";
 import KI0 from "../../appEtc/KI0.js";
 
 export default class Level21 extends Level {
