+++ datafiles/javascript/rooms/levels/Level25.js
@@ -1,5 +1,5 @@
 import Level from "./Level.js";
-import Bubble from "../../objects/Bubble.js";
+import Bubble from "../../objects/bases/Bubble.js";
 import KI2 from "../../appEtc/KI2.js";
 
 export default class Level25 extends Level {
