+++ datafiles/javascript/rooms/levels/Level33.js
@@ -1,5 +1,5 @@
 import Level from "./Level.js";
-import BubbleTemp from "../../objects/BubbleTemp.js";
+import BubbleTemp from "../../objects/bases/BubbleTemp/BubbleTemp.js";
 import KI0 from "../../appEtc/KI0.js";
 import ModFleeTemp from "../../appEtc/aiModules/ModFleeTemp.js";
